+++ src/main/java/com/leesang/mylocaldiary/member/repository/MemberRepository.java
@@ -8,4 +8,6 @@ import java.util.Optional;
 public interface MemberRepository extends JpaRepository<MemberEntity, Integer> {
 
     Optional<Object> findByLoginId(String loginId);
+
+    Optional<MemberEntity> findByEmail(String userEmail);
 }
