--- src/main/java/com/leesang/mylocaldiary/security/kakao/KakaoLoginController.java
@@ -1,31 +1,0 @@
-package com.leesang.mylocaldiary.security.kakao;
-
-import lombok.RequiredArgsConstructor;
-import lombok.extern.slf4j.Slf4j;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RequestParam;
-import org.springframework.web.bind.annotation.RestController;
-
-@Slf4j
-@RestController
-@RequiredArgsConstructor
-@RequestMapping("")
-public class KakaoLoginController {
-
-    private final KakaoService kakaoService;
-
-    @GetMapping("/callback")
-    public ResponseEntity<KakaoUserInfoResponseDto> callback(@RequestParam("code") String code) {
-        String accessToken = kakaoService.getAccessTokenFromKakao(code);
-
-        KakaoUserInfoResponseDto userInfo = kakaoService.getUserInfo(accessToken);
-
-        System.out.println("userInfo = " + userInfo);
-
-        // 여기에 서버 사용자 로그인(인증) 또는 회원가입 로직 추가
-        return ResponseEntity.ok(userInfo);
-    }
-}
