+++ src/main/java/com/leesang/mylocaldiary/security/kakao/controller/KakaoLoginPageController.java
@@ -1,4 +1,4 @@
-package com.leesang.mylocaldiary.security.kakao;
+package com.leesang.mylocaldiary.security.kakao.controller;
 
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Controller;
