+++ src/main/java/com/leesang/mylocaldiary/security/kakao/dto/KakaoUserInfoResponseDto.java
@@ -1,4 +1,4 @@
-package com.leesang.mylocaldiary.security.kakao;
+package com.leesang.mylocaldiary.security.kakao.dto;
 
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import com.fasterxml.jackson.annotation.JsonProperty;
