+++ src/main/resources/secret.properties.example
@@ -3,3 +3,4 @@ DB_PASSWORD=비밀번호
 S3_ACCESS_KEY=액세스키
 S3_SECRET_ACCESS_KEY=비밀액세스키
 REST_API_KEY=카카오
+secret_key
